Refactored nuget.exe pack icon tests
